pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Job search automation pipeline — scrape, score, and apply."
requires-python = ">=3.14"
dependencies = [
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings[yaml]>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-jobspy>=1.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",
    "rapidfuzz>=3.14",
"pymupdf4llm>=0.2.9",
    "weasyprint>=68.0",
    "sse-starlette>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pyright>=1.1.0",
    "pytest>=8.0.0",
    "factory-boy>=3.3.0",
    "Faker>=40.0.0",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "pytest-socket>=0.7.0",
    "pytest-cov>=7.0.0",
    "pytest-playwright>=0.6.0",
]

[project.scripts]
jobs-scrape = "core.orchestrator:main"
jobs-web = "webapp.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "platforms", "webapp", "resume_ai", "claude_cli", "apply_engine"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
markers = [
    "unit: Pure logic tests with no I/O",
    "integration: Tests that touch the database or combine modules",
    "e2e: End-to-end browser tests (requires Playwright)",
    "slow: Tests that take more than 5 seconds",
]
addopts = [
    "--disable-socket",
    "--allow-unix-socket",
    "-m", "not e2e",
    "--strict-markers",
]

[tool.coverage.run]
source = [
    "core",
    "platforms",
    "webapp",
    "resume_ai",
    "apply_engine",
    "claude_cli",
]
omit = [
    "*/test_*",
    "*/conftest*",
    "*/__pycache__/*",
    "platforms/stealth.py",
    "platforms/indeed.py",
    "platforms/dice.py",
    "platforms/indeed_selectors.py",
    "platforms/dice_selectors.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "pass",
]
